refactor(config): split config I/O
