+++ config/max-lines-baseline.txt
@@ -762,7 +762,6 @@ src/config/config.plugin-validation.test.ts
 src/config/env-preserve.ts
 src/config/io.observe-recovery.test.ts
 src/config/io.observe-recovery.ts
-src/config/io.ts
 src/config/io.write-config.test.ts
 src/config/io.write-prepare.test.ts
 src/config/io.write-prepare.ts
