+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-056f7cbfe3c61a7d9cfbb1cd0e026ecc570c56b9b0ea6a017ad8afa6be4cd5d5  plugin-sdk-api-baseline.json
-74d2c1c1be13364e8546a3be6bcaea541466249873d1f2d1085c793d55892cd6  plugin-sdk-api-baseline.jsonl
+153a0c98a554bb02e928fc3cb4e63d2d68cb13ba5e6f5f6a820f95c4b4c54e69  plugin-sdk-api-baseline.json
+8b47ed06bfced8f9785b15868b0367b3c5696c59928e430560080901d94c5bcd  plugin-sdk-api-baseline.jsonl
