+++ src/config/io.state.ts
@@ -1,0 +1,4 @@
+export const loggedInvalidConfigs = new Set<string>();
+export const loggedConfigWarningFingerprints = new Map<string, string>();
+export const warnedFutureTouchedVersions = new Set<string>();
+export const autoOwnerDisplaySecretByPath = new Map<string, string>();
